feat(theme-basic): Separate submit logic into plugins
The submit logic is no longer in the theme, but in plugins (now vote-config.js).
This is helpful to 'Don't Repeat Yourself', as submitting is a common logic.
Also, this commit adds a computed prop of Form: Form#formdata, in order to get the data of the form in a quick way.